Upgrade pole model (#1133)
* Create mdb-ami2glm-load.py

* Add sump pump to enduse loads

* Add sump state implementation

* Add sump enduse loadshape

* Revert "Add sump enduse loadshape"

This reverts commit 404995f6a0120df9a37bc56bfc37322716fa9b14.

* Create test_sump.glm

* Add sump to enduse schedules

* Add curtailable enduses global variable

* Add curtailment control global

* Add loadshape output command

* Add --loadshape command line option

* Add test_sump.glm

* Fix python setup

* Fix curtailment

* Fix linux build warnings

* Update cmdarg.cpp

* Disable dishwasher validation test because it doesn't work

* Fix enduse loadshape validation tests

* Update house model sump implementation

* Update billing.csv

* Update curtailment and sump docs and tests

* Create mdb-ami2glm-load.py

* Add general MDB converters

* Create mdb_info.py

* Update Makefile.mk

* Update mdb_info.py

* Create rbsa_single_family_2011.mdb

* Create ceus.glm

* Create config.glm

* Create ica_test.glm

* Create modify.glm

* Create build.h

* Create config.h

* Create config.h.in

* Create config.h.in~

* Create IEEE-123-cyme.png

* Create IEEE-13-cyme.png

* Create solver_nr_profile.csv

* Create table2glm_input.glm

* Create table2glm_input_noclass.glm

* Create table2glm_input_noname.glm

* Create gridlabd

* Create stamp-h1

* Update pole.cpp

* Update pole.h

* Update pole_configuration.cpp

* Update pole_mount.cpp

* Update pole_mount.h

* Update test_pole.csv

* Update Makefile.mk

* Update requirements.txt

* Update requirements.txt

* Update mdb_info.py

* MDB2GLM work progress

* Finish mdb-table to glm-object converter

* Update mdb-table2csv-player.py

* Add covered conductor

* Update load.cpp

* Work in progress

* Delete ami_data.mdb.gz

* Fix access to test data

* Update player.cpp

* Update load.cpp

* Update .gitignore

* Update load.cpp

* Update mdb-table2glm-player.py

* Create IEEE-13-cyme.glm

* Update test_mdbtable2glmobject.glm

* Create test_mdbtable2glmplayer.csv

* Update test_mdbtable2glmplayer.glm

* Update .gitignore

* Delete ami_data.glm

* Update IEEE-13-cyme.glm

* Update test_mdbtable2glmobject.glm

* Fix resolver

* Fix object name resolution

* Update test_create_poles.glm

* Fix requirements

* Update metrics.cpp

* Fix mdbtools usage in converters

* Update test_mdbtable2glmobject.glm

* Update xls-spida2csv-geodata.py

* Update test_xls-spida2csv-geodata.csv

* Update test_xls-spida2csv-geodata.csv

* Update json2html.py

* Update json2html_config.py

* Update xls-spida2csv-geodata.py

* Update Makefile.am

* Update pole_mount.cpp

* Update Makefile.am

* Update test_pole.csv

* Update create_poles.py

* Update noaa_forecast.py

* Update pole.cpp

+++ tools/create_poles.py
@@ -429,7 +429,7 @@ def main(inputfile,**options):
     for name in list(objects.keys()):
         data = model["objects"][name]
         if "from" in data.keys() and "to" in data.keys() and \
-                ( not "length" in data.keys() or int(data["length"].split()[0]) < short_line_length ):
+                ( not "length" in data.keys() or float(data["length"].split()[0]) < short_line_length ):
             fromname = data["from"]
             toname = data["to"]
